Refactored code to using single exception handling decorator
